requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
python-dotenv==1.0.1
httpx==0.27.2
pymupdf==1.24.10
pytesseract==0.3.13
Pillow==10.4.0
pydub==0.25.1
